Backend/requirements.txt
flask
flask-cors
torch
gensim
google-generativeai
grpcio
nltk
openai
pdfplumber
numpy
